refactor: rate -> review 네이밍 변경
